+++ src/NRelay1.test.ts
@@ -2,20 +2,28 @@ import { assert, assertEquals } from '@std/assert';
 import { finalizeEvent, generateSecretKey } from 'nostr-tools';
 
 import { NostrEvent } from '../interfaces/NostrEvent.ts';
-
+import { MockRelayWs } from './test/MockRelayWs.ts';
 import { NRelay1 } from './NRelay1.ts';
+import events from '../fixtures/events.json' with { type: 'json' };
+
+const event1s = events
+  .filter((e) => e.kind === 1)
+  .toSorted((_) => 0.5 - Math.random())
+  .slice(0, 10);
 
 Deno.test('NRelay1.query', async () => {
   const controller = new AbortController();
   const tid = setTimeout(() => controller.abort(), 3000);
 
-  const relay = new NRelay1('wss://relay.primal.net/');
+  const server = new MockRelayWs('wss://mock.relay', event1s);
+  const relay = new NRelay1('wss://mock.relay/');
   const events = await relay.query([{ kinds: [1], limit: 3 }], { signal: controller.signal });
 
   assertEquals(events.length, 3);
   assert(events[0].created_at >= events[1].created_at);
 
   await relay.close();
+  server.close();
   clearTimeout(tid);
 });
 
@@ -23,7 +31,8 @@ Deno.test('NRelay1.req', async () => {
   const controller = new AbortController();
   const tid = setTimeout(() => controller.abort(), 3000);
 
-  const relay = new NRelay1('wss://relay.primal.net/');
+  const server = new MockRelayWs('wss://mock.relay', event1s);
+  const relay = new NRelay1('wss://mock.relay/');
   const events: NostrEvent[] = [];
 
   for await (const msg of relay.req([{ kinds: [1], limit: 3 }], { signal: controller.signal })) {
@@ -36,14 +45,13 @@ Deno.test('NRelay1.req', async () => {
   assertEquals(events.length, 1);
 
   await relay.close();
+  server.close();
   clearTimeout(tid);
 });
 
 Deno.test('NRelay1.event', async () => {
-  const controller = new AbortController();
-  const tid = setTimeout(() => controller.abort(), 3000);
-
-  const relay = new NRelay1('wss://relay.mostr.pub/');
+  const server = new MockRelayWs('wss://mock.relay');
+  const relay = new NRelay1('wss://mock.relay/');
 
   const event: NostrEvent = finalizeEvent({
     kind: 1,
@@ -52,8 +60,7 @@ Deno.test('NRelay1.event', async () => {
     created_at: Math.floor(Date.now() / 1000),
   }, generateSecretKey());
 
-  await relay.event(event, { signal: controller.signal });
-
+  await relay.event(event);
   await relay.close();
-  clearTimeout(tid);
+  server.close();
 });